+++ examples/sveltekit/.svelte-kit/output/client/_app/immutable/nodes/0.BnJm1Jdz.js
@@ -1,1 +1,0 @@
-import{M as e,d as t,j as n,n as r,s as i,u as a,w as o}from"../chunks/DvTkoxpS.js";import"../chunks/xihTtKlq.js";function s(s,c){e(c,!0),r(async()=>{});var l=t();i(o(l),()=>c.children),a(s,l),n()}export{s as component};