+++ packages/next-pages/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @agent-devtools/next-pages
 
+## 0.7.0-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @agent-devtools/core@0.7.0-beta.0
+  - @agent-devtools/react@0.7.0-beta.0
+
 ## 0.6.0
 
 ### Minor Changes