+++ packages/widget-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @agent-devtools/widget-core
 
+## 0.7.0-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @agent-devtools/core@0.7.0-beta.0
+
 ## 0.6.0
 
 ### Minor Changes